requirements.txt
pandas
pymongo==4.6.2
python-dotenv==1.1.0
Flask==3.1.1
cohere==5.15.0
langchain==0.3.25
langchain_core==0.3.60
langchain-community==0.3.24
langchain-cohere==0.4.4
faiss-cpu==1.11.0